// src/app/administration/pages/appraiser/interfaces/valuator.checklist.interface.ts
export interface ValuatorChecklist {
    code:       string | number;
    status:     string;
    Check_List: CheckList;
}

export interface CheckList {
    current_page:   number;
    data:           Data[];
    first_page_url: string;
    from:           number;
    last_page:      number;
    last_page_url:  string;
    links:          Link[];
    next_page_url:  null;
    path:           string;
    per_page:       number;
    prev_page_url:  null;
    to:             number;
    total:          number;
}

export interface Data {
    id:                number;
    distributor:       string;
    valuation_date:    null;
    warranty_manual:   null;
    direct_purchase:   null;
    take_into_account: null;
    valid_warranty:    null;
    color:             null;
    cilindres:         null;
    plates:            null;
    origin_country:    null;
    transmission:      null;
    engine_suction:    null;
    start_stop:        null;
    /* req1:              null;
    req2:              null;
    req3:              null;
    req4:              null;
    req5:              null;
    req6:              null;
    req7:              null;
    req8:              null;
    req9:              null;
    req10:             null;
    req11:             null;
    req12:             null;
    req13:             null;
    req14:             null;
    req15:             null;
    req16:             null;
    req17:             null;
    req18:             null;
    req19:             null;
    req20:             null;
    req21:             null;
    req22:             null;
    iq1:               null;
    iq2:               null;
    iq3:               null;
    iq4:               null;
    iq5:               null;
    iq6:               null;
    iq7:               null;
    iq8:               null;
    iq9:               null;
    iq10:              null;
    iq11:              null;
    iq12:              null;
    iq13:              null;
    iq14:              null;
    iq15:              null;
    iq16:              null;
    iq17:              null;
    meq1:              null;
    meq2:              null;
    meq3:              null;
    meq4:              null;
    meq5:              null;
    meq6:              null;
    meq7:              null;
    meq8:              null;
    meq9:              null;
    meq10:             null;
    meq11:             null;
    meq12:             null;
    meq13:             null;
    meq14:             null;
    meq15:             null;
    meq16:             null;
    meq17:             null;
    meq18:             null;
    meq19:             null;
    meq20:             null;
    meq21:             null;
    meq22:             null;
    meq23:             null;
    meq24:             null;
    meq25:             null;
    meq26:             null;
    meq27:             null;
    meq28:             null;
    meq29:             null;
    meq30:             null;
    meq31:             null;
    meq32:             null;
    meq33:             null;
    meq34:             null;
    meq35:             null;
    meq36:             null;
    meq37:             null;
    breakedd:          null;
    breakeid:          null;
    breakeit:          null;
    breakedt:          null;
    meq38:             null;
    meq39:             null;
    meq40:             null;
    depthdd:           null;
    depthid:           null;
    depthit:           null;
    depthdt:           null;
    meq41:             null;
    meq42:             null;
    meq43:             null;
    meq44:             null;
    meq45:             null;
    meq46:             null;
    meq47:             null;
    meq48:             null;
    meq49:             null;
    meq50:             null;
    meq51:             null;
    cvq1:              null;
    cvq2:              null;
    cvq3:              null;
    cvq4:              null;
    cvq5:              null;
    cvq6:              null;
    cvq7:              null;
    cvq8:              null;
    cvq9:              null;
    cvq11:             null;
    cvq12:             null; */
    take:              null;
    sale:              null;
    workforce:         null;
    spare_parts:       null;
    hyp:               null;
    total:             null;
    take_value:        null;
    final_offer:       null;
    comments:          null;
    name_technical:    null;
    firm_technical:    null;
    name_manager:      null;
    firm_manager:      null;
    name_appraiser:    null;
    firm_appraiser:    null;
    status:            null;
    preparation:       null;
    technician_id:     null;
    sell_your_car_id:  number;
    created_at:        Date;
    updated_at:        Date;
}

export interface Link {
    url:    null | string;
    label:  string;
    active: boolean;
}