// src/app/administration/interfaces/mechanical-electric.interface.ts
export interface MechanicalElectric {
    status:                 string;
    code:                   string;
    DataMechanicElectronic: DataMechanicElectronic;
}

export interface DataMechanicElectronic {
    id:               number;
    meq1:             string;
    meq2:             string;
    meq3:             string;
    meq4:             string;
    meq5:             string;
    meq6:             string;
    meq7:             string;
    meq8:             string;
    meq9:             string;
    meq10:            string;
    meq11:            string;
    meq12:            string;
    meq13:            string;
    meq14:            string;
    meq15:            string;
    meq16:            string;
    meq17:            string;
    meq18:            string;
    meq19:            string;
    meq20:            string;
    meq21:            string;
    meq22:            string;
    meq23:            string;
    meq24:            string;
    meq25:            string;
    meq26:            string;
    meq27:            string;
    meq28:            string;
    meq29:            string;
    meq30:            string;
    meq31:            string;
    meq32:            string;
    meq33:            string;
    meq34:            string;
    meq35:            string;
    meq36:            string;
    meq37:            string;
    breakedd:         number;
    breakeid:         number;
    breakeit:         number;
    breakedt:         number;
    meq38:            string;
    meq39:            string;
    meq40:            string;
    depthdd:          number;
    depthid:          number;
    depthit:          number;
    depthdt:          number;
    meq41:            string;
    meq42:            string;
    meq43:            string;
    meq44:            string;
    meq45:            string;
    meq46:            string;
    meq47:            string;
    meq48:            string;
    meq49:            string;
    meq50:            string;
    meq51:            string;
    commentaryMechanical: string | null;
    sell_your_car_id: number;
    created_at:       Date;
    updated_at:       Date;
}